Clase partícula y suma a lo largo de los estados finales. Añadido: Clase particula con atributos int J, int M y vector h de probabilidad de cada estado de helicidad Función decaimiento que toma como argumentos la particula, el nombre del archivo de salida y el número de puntos Archivos de ejemplos del autor para comparar resultados si evtgen no funciona. NO ESTÄN en el makefile, pero la rutina de gráficos está como comentario en 3dplot.py Eliminado: Función iteradora a lo largo de theta y phi en Amplitudes -> main; ahora en Amplitudes -> decaimiento Sugerencias: Hacer una base de datos de partículas en Amplitudes.cpp Eliminar declaraciones no usadas de las cantidades en Amplitudes.cpp Hacer el archivo de gráficas en coordenadas polares

+++ 3dplot.py
@@ -6,14 +6,25 @@ ax = fig.add_subplot(projection='3d')
 ax.set_xlabel(r'$\theta$')
 ax.set_ylabel(r'$\phi$')
 ax.set_zlabel("Densidad de probabilidad")
-datos=np.genfromtxt("datos.dat",delimiter=",",skip_header=1)
+datos=np.genfromtxt("datos.dat",delimiter=",",skip_header=1,dtype=np.double)
 x=datos[:,0]
 y=datos[:,1]
 z=datos[:,2]
 ax.plot_trisurf(x, y, z)
 ax.view_init(25, 75)
-#Se muestra la distribución porque se pierde la capacidad de rotarla al guardar.
-#ERROR IMPORTANTE: LA GRÁFICA SIEMPRE SALE VACÍA. CORREGIR
 plt.savefig("Distribucion.pdf")
 plt.show()
 plt.close()
+
+#datos2=np.genfromtxt("WDecay.dat",delimiter=",",skip_header=0,dtype=np.double)
+#fig, ax = plt.subplots()
+
+#ax.plot(datos2[:,0],datos2[:,1])
+#plt.savefig("WDecay.pdf")
+#plt.show()
+#datos3=np.genfromtxt("JPsiDecay.dat",delimiter=",",skip_header=0,dtype=np.double)
+#fig, ax = plt.subplots()
+
+#ax.plot(datos3[:,0],datos3[:,1])
+#plt.savefig("JPsiDecay.pdf")
+#plt.show()
